requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.11
pymysql>=1.1.0
requests>=2.31.0
httpx>=0.25.2 

mysql-connector-python>=8.1.0

apscheduler>=3.10.4

python-dotenv>=1.0.0

pydantic-settings>=2.1.0  


python-dateutil>=2.8.2  

python-multipart>=0.0.6  